python-worker/requirements.txt
# =====================================================
# AI Resume Processing Worker - Dependencies (Fixed & Stable)
# System Requirements:
 Python==3.11.9
 Node==v24.13.0
 PostgreSQL==17.7
 Tesseract-OCR==5.5.0
# =====================================================

# -------------------------
# Core Framework
# -------------------------
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.12.5
pydantic-settings==2.1.0
python-multipart==0.0.6

# -------------------------
# PDF and Document Processing
# -------------------------
PyMuPDF==1.26.7
python-docx==0.8.11
pytesseract==0.3.10
Pillow==10.4.0
opencv-python==4.11.0.86
rapidfuzz==3.14.3

# -------------------------
# NLP & Machine Learning (Gemini API Based)
# -------------------------
# Using External LLM (Gemini) for both Parsing and Embeddings
# Local AI libraries (Torch/Sentence-Transformers) removed to save space
numpy==1.24.3

# -------------------------
# Database & Vector Search
# -------------------------
psycopg2-binary==2.9.9
asyncpg==0.29.0
pgvector==0.2.3

# -------------------------
# Message Queue (RabbitMQ / Redis)
# -------------------------
pika==1.3.2
celery==5.3.4
redis==5.0.1

# -------------------------
# Object Storage (MinIO / S3)
# -------------------------
minio==7.2.0
boto3==1.29.7

# -------------------------
# Utilities & Tools
# -------------------------
python-dotenv==1.0.0
requests==2.31.0
aiofiles==23.2.1
loguru==0.7.2

# -------------------------
# External LLM (Gemini)
# -------------------------
google-generativeai==0.8.3

# -------------------------
# Data Processing
# -------------------------
pandas==2.1.3
scikit-learn==1.3.2
openpyxl==3.1.2
xlrd==2.0.1

# -------------------------
# Logging & Monitoring
# -------------------------
prometheus-client==0.19.0

# -------------------------
# Testing & Debugging
# -------------------------
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
